Declare Prophesee formats to build on vanilla kernel
It is possible to have the format definition in the kernel module
and in the userland, without having it in the kernel headers, allowing
to use this kernel module without patching and rebuilding the kernel.
It rises some warnings, and shouldn't be the preferred way to use this,
but it sometimes helps for quick hacks.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>