Probe psee-csi2rxss only if VFB is not present
Testing the opposite condition with regard to the xlnx driver, this
allow to have both driver in the image in case we want both an
event-based and a frame-based acquisition pipeline.

Formally, I should also remove the part where the code checks if the
requested data type is the one accepted by the VFB, but since we use
user defined types that always go through, I leave that as en exercise
for the reader.

Overall, I expect everything to fit in a single driver, but let's first
have a working demo, and then get in touch with AMD to merge everything.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>