Add bindings for psee,axi4s-packetizer
There is room to add additional ports on the IP, making the current
binding a little weird, but that way current bindings will still work
with an upgraded schema version.

The clocks are missing in this version. A name need to be chosen,
because, at some point, the driver will need it to configure timeouts.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>