Apply ps_host_if options to all 2.x versions
The strict test was meant to handle an incoming change int the fields
of the version register, but We are going to change the register maps
of every IP, including the location of the version register, and drop
support of the old versions, so there's no need to be so pedantic. In
the mean time, there are some fixes to deploy for some demos.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>